+++ pages/ingredientes.vue
@@ -57,6 +57,7 @@
 
 <script setup>
 import { ref, onMounted } from 'vue'
+import Swal from 'sweetalert2';
 
 const { token } = useAuth()
 
@@ -139,18 +140,45 @@ const actualizarIngrediente = async () => {
 }
 
 const eliminarIngrediente = async (id) => {
-  try {
-    await $fetch(`http://localhost:3001/ingredientes/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Authorization': token.value
-      }
-    })
-    await obtenerIngredientes()
-  } catch (error) {
-    console.error('Error al eliminar ingrediente:', error)
+  const result = await Swal.fire({
+    title: '¿Estás seguro?',
+    text: '¡No podrás revertir esto!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Sí, eliminar',
+    cancelButtonText: 'Cancelar',
+  });
+
+  if (result.isConfirmed) {
+    try {
+      await $fetch(`http://localhost:3001/ingredientes/${id}`, {
+        method: 'DELETE',
+        headers: {
+          'Authorization': token.value
+        }
+      });
+      await obtenerIngredientes();
+
+      // Mostrar mensaje de éxito
+      Swal.fire(
+        '¡Eliminado!',
+        'El ingrediente ha sido eliminado.',
+        'success'
+      );
+    } catch (error) {
+      console.error('Error al eliminar ingrediente:', error);
+
+      // Mostrar mensaje de error
+      Swal.fire(
+        'Error',
+        'No se pudo eliminar el ingrediente.',
+        'error'
+      );
+    }
   }
-}
+};
 
 const limpiarFormulario = () => {
   ingredienteId.value = null
